fix(zsh): move environment setup to zshenv in home directory

+++ config/zsh/.zshrc
@@ -1,25 +1,28 @@
-[ -f "/opt/homebrew/bin/brew" ] && eval "$(/opt/homebrew/bin/brew shellenv)"
-
-# Path
-path+=($HOME/.local/bin)
+# MacOS specific
+if [[ "$OSTYPE" == "darwin"* ]]; then
+	[ -f "/opt/homebrew/bin/brew" ] && eval "$(/opt/homebrew/bin/brew shellenv)"
+	# Remove git completions from brew in favor of the zsh ones
+	rm -f "$HOMEBREW_PREFIX/share/zsh/site-functions/_git"
+fi
 
 # Plugins
 eval "$(sheldon source)"
 source <(fzf --zsh)
 
-# Load and initialise completion system
-autoload -Uz compinit && compinit
-# Remove git completions from brew in favor of the zsh ones
-rm -f "$HOMEBREW_PREFIX/share/zsh/site-functions/_git"
 
 # History
-HISTFILE="$XDG_STATE_HOME/zsh/history"
+mkdir -p "${XDG_STATE_HOME:-$HOME/.local/state}/zsh"
+HISTFILE="${XDG_STATE_HOME:-$HOME/.local/state}/zsh/history"
 HISTSIZE=10000000
 SAVEHIST=10000000
 setopt HIST_IGNORE_DUPS
 setopt HIST_FIND_NO_DUPS
 setopt INC_APPEND_HISTORY
 setopt HIST_IGNORE_SPACE
+
+
+# Completions
+autoload -Uz compinit && compinit
 
 zstyle ':completion:*' menu select
 zstyle ':completion:*' matcher-list 'm:{a-zA-Z}={A-Za-z}' 'r:|[._-]=* r:|=*' 'l:|=* r:|=*'
@@ -34,12 +37,16 @@ zstyle ':completion:*' group-name ''
 zstyle ':completion:*:*:-command-:*:*' group-order aliases builtins functions commands
 zstyle ':completion:*:*:cd:*' tag-order local-directories directory-stack path-directories
 
+
+# Misc
 setopt MENU_COMPLETE # Automatically select first option
 setopt AUTO_LIST     # Show list on first Tab press
 setopt NO_LIST_BEEP  # Prevent beeps when showing list
 
+
 # Ensure emcas mode
 set -o emacs
+
 
 # Make all PATH entires unique
 typeset -aU path
